Refactor the code by Omar and Kerolous to arrange the modules
